requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
ormar==0.12.2
alembic==1.12.0
psycopg2-binary==2.9.9
asyncpg==0.28.0 
python-dotenv==1.0.0
